requirements.txt
altgraph==0.17
pefile==2019.4.18
Pillow==8.1.0
psutil==5.8.0
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.11
pywin32-ctypes==0.2.0
selenium==3.141.0
urllib3==1.26.2
